Rename old thrift
