+++ src/meta/upgrade/CMakeLists.txt
@@ -1,7 +1,7 @@
 nebula_add_library(
     meta_data_upgrade_obj OBJECT
     MetaDataUpgrade.cpp
-    oldThrift/MetaServiceUtilsV1.cpp
+    thrift/MetaServiceUtilsV1.cpp
 )
 
 add_dependencies(
@@ -9,4 +9,4 @@ add_dependencies(
     old_meta_thrift_obj
 )
 
-add_subdirectory(oldThrift)
+add_subdirectory(thrift)
