+++ src/meta/upgrade/MetaDataUpgrade.h
@@ -11,7 +11,7 @@
 #include "common/base/Base.h"
 #include "common/base/Status.h"
 #include "interface/gen-cpp2/meta_types.h"
-#include "meta/upgrade/oldThrift/gen-cpp2/old_meta_types.h"
+#include "meta/upgrade/thrift/gen-cpp2/old_meta_types.h"
 
 
 namespace nebula {
